% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_s3methods.R
\name{as.matrix.FastTransform}
\alias{as.matrix.FastTransform}
\title{Convert fast transform to matrix}
\usage{
\method{as.matrix}{FastTransform}(x, ...)
}
\arguments{
\item{x}{An object of class \code{FastTransform}, created using \code{fort()}.}

\item{...}{Extra parameters (ignored).}
}
\value{
A \code{matrix} object equivalent to x.
}
\description{
Converts a fast transform created by \code{fort()} to the equivalent matrix form.
}
\examples{
fast_transform <- fort(4,15)
slow_transform <- as.matrix(fast_transform)
fast_result <- fast_transform \%*\% diag(4)
slow_result <- slow_transform \%*\% diag(4)
norm(fast_result - slow_result) # should be small
}
