% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projector.R
\name{Cprojections}
\alias{Cprojections}
\title{Cprojections}
\usage{
Cprojections(
  year,
  Ihat,
  sEI,
  Nhat,
  sEN,
  Mhat,
  sEM,
  Phat,
  sEP,
  logIRR,
  logIRRdelta,
  ...,
  nahead = 0,
  returntype = "projection",
  modeltype = "rwI",
  verbose = FALSE
)
}
\arguments{
\item{year}{vector of years for which there is data or an estimate is required}

\item{Ihat}{Incidence midpoints (NA if projection/imputation needed)}

\item{sEI}{Incidence uncertainty as SD (NA if projection/imputation needed)}

\item{Nhat}{Notifications midpoints (NA if projection/imputation needed)}

\item{sEN}{Notifications uncertainty as SD (NA if projection/imputation needed)}

\item{Mhat}{Untreated Deaths midpoints (NA if projection/imputation needed)}

\item{sEM}{Untreated Deaths uncertainty as SD (NA if projection/imputation needed)}

\item{Phat}{Prevalence midpoints (NA if projection/imputation needed)}

\item{sEP}{Prevalence uncertainty as SD (NA if projection/imputation needed)}

\item{logIRR}{Log-hazard ratios to apply to incidence (assumed 1 if not given)}

\item{logIRRdelta}{Log-hazard ratios to apply to detection hazard (assumed 1 if not given)}

\item{...}{Additional parameters to help tune away from defaults}

\item{nahead}{how many years to project
\itemize{
\item futureonly: (default) only returns projection
\item fit: returns SSM output for all times with data
\item projection: returns inputs during data combined with projections after
\item projectionfit: combination of above
}}

\item{returntype}{Determines what is returned if projecting}

\item{modeltype}{String to specify which SSM variant to use:
\itemize{
\item rwI: (default) random walk for incidence. NOTE no indirect impact
\item IP: AR(1) on a mixture of log(Incidence) and log(Prevalence).
\item IP(n)k: AR(1) as above, but \loadmathjax\mjeqn{I_{t+1} = R\times I_t^(1-\mathrm{expit}(k))P_t^{\mathrm{expit}(k)} }{ascii}  with k=-k with n present. IP = IP0
}}

\item{verbose}{(Default=FALSE) Give more output for use in debugging.}
}
\value{
A data.frame/data.table with the projections and uncertainty
}
\description{
SSM-based Projections
}
\details{
This wraps the SSM-based models and projections using BSSM.
}
\examples{

1+1 #TODO

}
\author{
Pete Dodd
}
