% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_s3methods.R
\name{dim.FastTransform}
\alias{dim.FastTransform}
\title{Dimensions of fast transform}
\usage{
\method{dim}{FastTransform}(x)
}
\arguments{
\item{x}{An object of class \code{FastTransform}, created using \code{fort()}.}
}
\value{
A vector of length 2 containing the dimensions of the fast transform (i.e., number of rows and
number of columns, in this order).
}
\description{
Retrieves the dimensions of a fast transform created by \code{fort()} (i.e., the number of rows and
columns of an equivalent matrix). It returns the same value that one would get from \code{dim(as.matrix())},
but much more efficiently.
}
\examples{
dim(fort(3, 17)) # should return c(17,3)
dim(t(fort(3, 17))) # should return c(3,17)
}
\seealso{
\code{\link[=fort]{fort()}}
}
