% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_class.R
\name{.is_valid_ft}
\alias{.is_valid_ft}
\title{Validate \code{FastTransform} Objects}
\usage{
.is_valid_ft(object)
}
\arguments{
\item{object}{Input object to validate.}
}
\value{
Logical with (character) 'reason' attribute.
}
\description{
Internal function used to verify whether a certain object is a valid and
usable object that inherits from \code{FastTransform}. Thorough, but only used
for tests.
}
