% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exparz}
\alias{exparz}
\title{exparz}
\usage{
exparz(mu, sdlog)
}
\arguments{
\item{mu}{lognormal distribution mu parameter}

\item{sdlog}{lognormal distribution sigma parameter}
}
\value{
list with mn (mean) and S (SD) of associated lognormal distribution
}
\description{
Lognormal Distribution Mean and SD
}
\details{
Compute the mean and SD for a lognormal distribution with given parameters
}
\author{
Pete Dodd
}
