% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expit}
\alias{expit}
\title{expit}
\usage{
expit(x)
}
\arguments{
\item{x}{a vector of inputs}
}
\value{
1/(1+exp(-x))
}
\description{
Inverse Logit Transformation
}
\details{
Taken from \code{logitnorm} package
}
\author{
Pete Dodd
}
