% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_class.R
\name{FastTransform}
\alias{FastTransform}
\title{\code{FastTransform} class}
\description{
General specification of the type of objects generated by \code{fort()}, which correspond
to structured linear transforms. Useful objects of this class must be also part of a subclass
which extends this one with a specific implementation (e.g., \code{FastTransformFFT2}).
}
\details{
It is generally \emph{not} recommended that the fields and methods described here are used
directly, unless you need to minimize validation overhead when using \emph{pre-validated} inputs.
Instead, you should use \code{fort()} and the typical S3 methods for matrices (\verb{\%*\%.FastTransform},
\verb{t().FastTransform}, \verb{solve().FastTransform}, etc.).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{inverse}}{Logical. Indicates whether the object currently represents a forward or
inverse transform.}

\item{\code{invertible}}{Logical. Indicates whether the inverse transform can also be expressed as a
\code{FastTransform} object.}

\item{\code{dim_in}}{Dimensionality of the input for the forward transform.}

\item{\code{dim_out}}{Dimensionality of the output for the forward transform.}

\item{\code{blocksize}}{Dimensionality of the internal transformation (always a power of 2).}

\item{\code{fwd_par}}{List of parameters used in the forward transform.}

\item{\code{fwd_mtrx}}{Cached matrix representation of the forward transform.}

\item{\code{rev_par}}{List of parameters used in the inverse transform.}

\item{\code{rev_mtrx}}{Cached matrix representation of the inverse transform.}

\item{\code{fort_type}}{String indicating the type of structured transform being used.}

\item{\code{cache_matrix}}{Logical. Indicates whether to cache calculated matrices or not (default is \code{TRUE}).}

\item{\code{logdet}}{List with cached determinants of the forward and inverse transforms}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FastTransform-fwd_eval}{\code{FastTransform$fwd_eval()}}
\item \href{#method-FastTransform-rev_eval}{\code{FastTransform$rev_eval()}}
\item \href{#method-FastTransform-calculate_rev_par}{\code{FastTransform$calculate_rev_par()}}
\item \href{#method-FastTransform-new}{\code{FastTransform$new()}}
\item \href{#method-FastTransform-evaluate}{\code{FastTransform$evaluate()}}
\item \href{#method-FastTransform-get_ncol}{\code{FastTransform$get_ncol()}}
\item \href{#method-FastTransform-get_nrow}{\code{FastTransform$get_nrow()}}
\item \href{#method-FastTransform-get_dim}{\code{FastTransform$get_dim()}}
\item \href{#method-FastTransform-get_n_par}{\code{FastTransform$get_n_par()}}
\item \href{#method-FastTransform-get_inverse}{\code{FastTransform$get_inverse()}}
\item \href{#method-FastTransform-get_transpose}{\code{FastTransform$get_transpose()}}
\item \href{#method-FastTransform-get_logdet}{\code{FastTransform$get_logdet()}}
\item \href{#method-FastTransform-get_norm}{\code{FastTransform$get_norm()}}
\item \href{#method-FastTransform-get_norm_margin}{\code{FastTransform$get_norm_margin()}}
\item \href{#method-FastTransform-as_matrix}{\code{FastTransform$as_matrix()}}
\item \href{#method-FastTransform-print}{\code{FastTransform$print()}}
\item \href{#method-FastTransform-summary}{\code{FastTransform$summary()}}
\item \href{#method-FastTransform-clone}{\code{FastTransform$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-fwd_eval"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-fwd_eval}{}}}
\subsection{Method \code{fwd_eval()}}{
Function that performs the forward transform.
Do \emph{not} call this directly unless you know what you are doing:
use the \code{FastTransform$evaluate()} method instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$fwd_eval(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Input matrix of the \emph{correct} dimensionality}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix with the same number of columns as \code{x}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-rev_eval"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-rev_eval}{}}}
\subsection{Method \code{rev_eval()}}{
Function that performs the inverse transform.
Do \emph{not} call this directly unless you know what you are doing:
use the \code{FastTransform$evaluate()} method instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$rev_eval(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Input matrix of the \emph{correct} dimensionality}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix with the same number of columns as \code{x}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-calculate_rev_par"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-calculate_rev_par}{}}}
\subsection{Method \code{calculate_rev_par()}}{
Function that calculates and caches the parameters
for the inverse transform. Do not call this directly unless you
know what you are doing. If you need the inverse transform, use
the \code{FastTransform$get_inverse()} method instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$calculate_rev_par()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-new"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-new}{}}}
\subsection{Method \code{new()}}{
Raw object creation function. Note that calling this
function does \emph{not} result in a useful object. Instead, you
should call the \code{fort()} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$new(dim_in, dim_out, blocksize)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dim_in}}{Dimensionality of the input for the forward transform.}

\item{\code{dim_out}}{Dimensionality of the output for the forward transform.}

\item{\code{blocksize}}{Dimensionality of the internal transformation
(\emph{must} be a power of 2).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix with the same number of columns as \code{x}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
Evaluates the result of applying the transform represented
by this object on an input matrix \code{x}. It is important that the
provided matrix has compatible dimensionality since \emph{no input
validation is performed}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$evaluate(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Input matrix with correct dimensionality.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix with the same number of columns as \code{x}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-get_ncol"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-get_ncol}{}}}
\subsection{Method \code{get_ncol()}}{
Returns the number of columns of the linear
transform represented by this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$get_ncol()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-get_nrow"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-get_nrow}{}}}
\subsection{Method \code{get_nrow()}}{
Returns the number of rows of the linear
transform represented by this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$get_nrow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-get_dim"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-get_dim}{}}}
\subsection{Method \code{get_dim()}}{
Returns the dimensions of the linear
transform represented by this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$get_dim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric vector with length 2.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-get_n_par"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-get_n_par}{}}}
\subsection{Method \code{get_n_par()}}{
Returns the number of parameters required to represent
the linear transform represented by this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$get_n_par()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-get_inverse"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-get_inverse}{}}}
\subsection{Method \code{get_inverse()}}{
Returns a new \code{FastTransform} object that represents
the inverse transform of the transform represented by this
object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$get_inverse()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new object of type \code{FastTransform}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-get_transpose"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-get_transpose}{}}}
\subsection{Method \code{get_transpose()}}{
Returns either a \code{FastTransform} object that represents
the transpose of the transform represented by this object (if
the value of the \code{invertible} field is \code{TRUE}), or an equivalent
matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$get_transpose()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Either an new object of type \code{FastTransform} or a matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-get_logdet"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-get_logdet}{}}}
\subsection{Method \code{get_logdet()}}{
Returns information on the determinant of the
transform represented by this object. Fails if \code{dim_in != dim_out}.
The modulus of the determinant is provided in log scale.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$get_logdet()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The same type of object returned by \code{determinant.matrix}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-get_norm"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-get_norm}{}}}
\subsection{Method \code{get_norm()}}{
Returns norm of the matrix equivalent to the linear
transform represented by this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$get_norm(type = "o")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{String indicating the type of matrix norm to calculate,
using the same convention as \code{base::norm} (default is "o", which
corresponds to the maximum absolute column sum).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-get_norm_margin"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-get_norm_margin}{}}}
\subsection{Method \code{get_norm_margin()}}{
Returns norms of the rows (or columns) or the matrix
equivalent to the linear transform represented by this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$get_norm_margin(type = "2", by = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{String indicating the type of matrix norm to calculate,
using a convention compatible with \code{base::norm} (default is "2",
which corresponds to the Euclidian norm; use "o" for L1 norm,
"m" for Inf norm).}

\item{\code{by}}{The norms of the rows are calculated by default (\code{by = 1}).
To calculate the norms of columns instead, use \code{by = 2}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of numeric values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-as_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-as_matrix}{}}}
\subsection{Method \code{as_matrix()}}{
Returns the matrix equivalent to the transform
represented by this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$as_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-print"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-print}{}}}
\subsection{Method \code{print()}}{
Prints terse information about the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The object itself (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-summary"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-summary}{}}}
\subsection{Method \code{summary()}}{
Prints verbose information about the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The object itself (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastTransform-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FastTransform-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastTransform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
