% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLNmu}
\alias{getLNmu}
\title{getLNmu}
\usage{
getLNmu(m, S)
}
\arguments{
\item{m}{mean of target}

\item{S}{SD of target}
}
\value{
mu for lognormal distribution
}
\description{
Parametrize Lognormal Distribution
}
\details{
Lognormal distribution parameters from moment-matching
}
\author{
Pete Dodd
}
