% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLNsig}
\alias{getLNsig}
\title{getLNsig}
\usage{
getLNsig(m, S)
}
\arguments{
\item{m}{mean of target}

\item{S}{SD of target}
}
\value{
sigma (sdlog) for lognormal distribution
}
\description{
Parametrize Lognormal Distribution
}
\details{
Lognormal distribution parameters from moment-matching
}
\author{
Pete Dodd
}
