% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_s3methods.R
\name{\%*\%.FastTransform}
\alias{\%*\%.FastTransform}
\title{Apply a fast transform}
\usage{
\method{\%*\%}{FastTransform}(x, y)
}
\arguments{
\item{x}{An object of class \code{FastTransform}, created using \code{fort()}.}

\item{y}{A numeric (real) matrix/vector with an appropriate number of rows/elements.}
}
\value{
A numeric (real) matrix with the same number of columns as \code{y}.
}
\description{
Applies a fast transform created by \code{fort()} (\code{x}) to the columns of a conformable matrix (\code{y}).
}
\examples{
Z <- fort(4,1024)
Z \%*\% matrix(1:2,4,3) # output is a 1024 by 3 matrix
# the example below works: y is assumed to be a single column vector
Z \%*\% 1:4 # output is a 1024 by 1 matrix
}
