% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_unsafe.R
\name{\%***\%}
\alias{\%***\%}
\title{Unsafely apply a fast transform}
\usage{
x \%***\% y
}
\arguments{
\item{x}{An object of class \code{FastTransform}, created using \code{fort()}.}

\item{y}{A numeric (real) matrix/vector with an appropriate number of rows/elements.}
}
\value{
A numeric (real) matrix with the same number of columns as \code{y}.
}
\description{
Applies a fast transform created by \code{fort()} (\code{x}) to the columns of a conformable matrix (\code{y}),
typically equivalent to the use of the \code{\%*\%} operator, but using an \strong{unsafe} method.
}
\details{
This operator works in a similar way to \code{\%*\%}, but avoids dispatching and does not perform any type of
validation of its inputs, in order to reduce overhead when performing repeated operations inside a
function on \emph{pre-validated} inputs.

It is \emph{not} recommended that this operator is used interactively and/or on non-validated inputs.
}
\examples{
Z <- fort(4, 1024)
Z \%*\% matrix(1:2, 4, 3) # output is a 1024 by 3 matrix
Z \%***\% matrix(1:2, 4, 3) # output is also a 1024 by 3 matrix
}
\seealso{
\code{\link[=fort]{fort()}}, \code{\link{\%*\%.FastTransform}}
}
