% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lgtparz}
\alias{lgtparz}
\title{lgtparz}
\usage{
lgtparz(mn, S)
}
\arguments{
\item{mn}{mean of target}

\item{S}{SD of target}
}
\value{
list with mu (mu) and sigma (sdlog) parameters for logitnormal distribution
}
\description{
Parametrize Logitnormal Distribution
}
\details{
Parametrize logitnormal distribution based on mean/SD. Wrapper for \code{logitnorm::twCoefLogitnormCi}
}
\author{
Pete Dodd
}
