% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lnparz}
\alias{lnparz}
\title{lnparz}
\usage{
lnparz(mn, S)
}
\arguments{
\item{mn}{mean of target}

\item{S}{SD of target}
}
\value{
list with mu (mu) and sigma (sdlog) for lognormal distribution
}
\description{
Parametrize Lognormal Distribution
}
\details{
Lognormal distribution parameters from moment-matching
}
\author{
Pete Dodd
}
