% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\title{logit}
\usage{
logit(x)
}
\arguments{
\item{x}{a vector of inputs}
}
\value{
logit(x)
}
\description{
Logit Transformation
}
\details{
Taken from \code{logitnorm} package
}
\author{
Pete Dodd
}
