% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projector.R
\name{mcmcsmry}
\alias{mcmcsmry}
\title{MCMC to summary}
\usage{
mcmcsmry(fit)
}
\arguments{
\item{fit}{an MCMC fit object from fitting with bssm}
}
\value{
a data.table with mid,lo,hi by time and variable
}
\description{
MCMC Summary Computer
}
\details{
Computes mid (mean) and hi/lo (95\% quantiles) by variable and time and also adds in exponentiated variables if their names begin with log.
}
\author{
Pete Dodd
}
