% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projector.R
\name{projections}
\alias{projections}
\title{projections}
\usage{
projections(year, Ihat, sEI, Nhat, sEN, Mhat, sEM, Phat, sEP, nrep = 500)
}
\arguments{
\item{year}{vector of years for which there is data or an estimate is required}

\item{Ihat}{Incidence midpoints (NA if projection/imputation needed)}

\item{sEI}{Incidence uncertainty as SD (NA if projection/imputation needed)}

\item{Nhat}{Notifications midpoints (NA if projection/imputation needed)}

\item{sEN}{Notifications uncertainty as SD (NA if projection/imputation needed)}

\item{Mhat}{Deaths midpoints (NA if projection/imputation needed)}

\item{sEM}{Deaths uncertainty as SD (NA if projection/imputation needed)}

\item{Phat}{Prevalence midpoints (NA if projection/imputation needed)}

\item{sEP}{Prevalence uncertainty as SD (NA if projection/imputation needed)}

\item{nrep}{number of replicates used - NOTE likely to be temporary}
}
\value{
A data.frame/data.table with the projections and uncertainty
}
\description{
Main projection function to be exported
}
\details{
.. content for \details{projections} ..
}
\author{
Pete Dodd
}
