% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_s3methods.R
\name{solve.FastTransform}
\alias{solve.FastTransform}
\title{Solve a System of Equations}
\usage{
\method{solve}{FastTransform}(a, b, ...)
}
\arguments{
\item{a}{An object of class \code{FastTransform}, created using \code{fort()}.}

\item{b}{A numeric vector or matrix (to solve the equation), or nothing (to obtain a generalized
inverse of \code{a}).}

\item{...}{Extra parameters (ignored).}
}
\value{
Either a matrix (representing \code{x}), or a \code{FastTransform} object (representing a generalized
inverse of \code{a}; if parameter \code{b} is missing).
}
\description{
Solves an equation of the form \code{a \%*\% x = b} for \code{x}, where \code{a} is a linear operation represented by
a \code{FastTransform} object, while \code{b} can be either a vector or a matrix. If \code{b} is missing, it returns
a \code{FastTransform} object corresponding to the inverse (or a generalized inverse) of \code{a}.
}
\details{
Note that the inverse transform will only be fast (i.e., avoid matrix multiplication) if
\eqn{dim\_in = dim\_out = blocksize}.
}
\examples{
a <- fort(4)
inv_a <- solve(a) # inverse of a
inv_a \%*\% diag(4) # applying the inverse of a
solve(a, diag(4)) # should give the same output
}
\seealso{
\code{\link[=fort]{fort()}}
}
