% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_s3methods.R
\name{summary.FastTransform}
\alias{summary.FastTransform}
\title{Summarize fast transform}
\usage{
\method{summary}{FastTransform}(object, ...)
}
\arguments{
\item{object}{An object of class \code{FastTransform}, created using \code{fort()}.}

\item{...}{Extra parameters (ignored).}
}
\value{
The input object (invisibly).
}
\description{
Provides a summary of a fast transform created by \code{fort()} with slightly more detail than the
information provided by using \code{print()}.
}
\examples{
summary(fort(3, 17))
}
\seealso{
\code{\link[=fort]{fort()}}
}
