% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_s3methods.R
\name{t.FastTransform}
\alias{t.FastTransform}
\title{Transform Transpose}
\usage{
\method{t}{FastTransform}(x)
}
\arguments{
\item{x}{An object of class \code{FastTransform}, created using \code{fort()}.}
}
\value{
Either an object of class \code{FastTransform} (if \code{x$invertible} is \code{TRUE}) or a matrix.
}
\description{
Given a \code{FastTransform} object \code{x}, \code{t} returns the transpose of \code{x}. If \code{x} represents an orthonormal
transformation (i.e., if \code{x$invertible} is \code{TRUE}), then a \code{FastTransform} object (representing
the transpose of \code{x}) will be returned; otherwise, a matrix object (representing the transpose of \code{x})
will be returned, with a warning.
}
\examples{
(a <- fort(4))
(b <- t(t(a))) # transpose a twice
# the result below should be close to zero
sum((a \%*\% diag(4) - b \%*\% diag(4))^2)
}
\seealso{
\code{\link[=fort]{fort()}} and \code{\link[=solve.FastTransform]{solve.FastTransform()}}
}
