% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projector.R
\name{Cprojections}
\alias{Cprojections}
\title{Cprojections}
\usage{
Cprojections(
  year,
  Ihat,
  sEI,
  Nhat,
  sEN,
  Mhat,
  sEM,
  Phat,
  sEP,
  ...,
  nahead = 0,
  returntype = "projection",
  modeltype = "rwI",
  verbose = FALSE
)
}
\arguments{
\item{year}{vector of years for which there is data or an estimate is required}

\item{Ihat}{Incidence midpoints (NA if projection/imputation needed)}

\item{sEI}{Incidence uncertainty as SD (NA if projection/imputation needed)}

\item{Nhat}{Notifications midpoints (NA if projection/imputation needed)}

\item{sEN}{Notifications uncertainty as SD (NA if projection/imputation needed)}

\item{Mhat}{Deaths midpoints (NA if projection/imputation needed)}

\item{sEM}{Deaths uncertainty as SD (NA if projection/imputation needed)}

\item{Phat}{Prevalence midpoints (NA if projection/imputation needed)}

\item{sEP}{Prevalence uncertainty as SD (NA if projection/imputation needed)}

\item{nahead}{how many years to project
\itemize{
\item `futureonly': (default) only returns projection
\item `fit': returns SSM output for all times
\item \code{projection}: returns inputs during data combined with projections after
}}

\item{returntype}{Determines what is returned if projecting}

\item{modeltype}{String to specify which SSM variant to use:
\itemize{
\item `rwI': (default) random walk for incidence. NOTE no indirect impact
\item `IP': AR(1) on a mixture of log(Incidence) and log(Prevalence).
}}

\item{verbose}{(Default=FALSE) Give more output for use in debugging.}
}
\value{
A data.frame/data.table with the projections and uncertainty
}
\description{
SSM-based Projections
}
\details{
This wraps the SSM-based models and projections using BSSM.
}
\examples{

1+1 #TODO

}
\author{
Pete Dodd
}
