% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fort-package.R
\docType{package}
\name{fort-package}
\alias{fort-package}
\title{fort: Fast Orthogonal Random Transforms in R}
\description{
Provides convenient access to fast structured random linear transforms implemented in C++ (via 'Rcpp') that are (at least approximately) orthogonal or semi-orthogonal, and often much faster than matrix multiplication. Useful for algorithms that require or benefit from uncorrelated random projections, such as fast dimensionality reduction (e.g., Johnson-Lindenstrauss transform) or kernel approximation (e.g., random kitchen sinks) methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tomessilva/fort}
  \item \url{https://tomessilva.github.io/fort/}
  \item Report bugs at \url{https://github.com/tomessilva/fort/issues}
}

}
\author{
\strong{Maintainer}: Tomé Silva \email{tome@tomesilva.com} (\href{https://orcid.org/0000-0002-9434-8686}{ORCID})

}
\keyword{internal}
