% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projector.R
\name{noisyex}
\alias{noisyex}
\title{Simulating and fitting replicates with structural model}
\usage{
noisyex(yrz, mnz, sdz, nrep, runs = TRUE, trnsfm = 0)
}
\arguments{
\item{yrz}{vector of years}

\item{mnz}{vector of means}

\item{sdz}{vector of SDs}

\item{nrep}{number of replicates to use}

\item{runs}{logical: whether to return individual run data}

\item{trnsfm}{0 = no transformation; 1 = log transformation; 2 = logit transformation}
}
\value{
data.table of results
}
\description{
Utility For Simulation-based Structural Time Series Fitting & Projection
}
\details{
Samples timeseries within uncertainty bounds and uses a Kalman filter from imputeTS to project.
}
\author{
Pete Dodd
}
