% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projector.R
\name{noisyex}
\alias{noisyex}
\title{Simulating and fitting replicates with structural model}
\usage{
noisyex(yrz, mnz, sdz, nrep, runs = TRUE, trnsfm = 0)
}
\arguments{
\item{yrz}{}

\item{mnz}{}

\item{sdz}{}

\item{nrep}{}

\item{runs}{}

\item{trnsfm}{0 = no transformation; 1 = log transformation; 2 = logit transformation}
}
\value{
data.table of results
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Pete Dodd
}
